tests/wpt/web-platform-tests/shadow-dom/focus-navigation/focus-navigation-scroller-delegatesFocus.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/declarative-shadow-dom-polyfill.js"></script>
<script src="resources/shadow-dom.js"></script>
<script src="resources/focus-utils.js"></script>

<style>
  div.scroll {
    overflow: auto;
    width: 20em;
    height: 5em;
    display: block;
  }
</style>

<p>Tests for moving focus on scrollable element by pressing tab key.</p>
<button id="start">START</button>

<p>Where host has delegatesFocus false.</p>
<div id="host-no-delegates-focus" class="scroll">
  <template shadowrootmode="open">
    <div style="width: 40em"></div>
    <button id="A">BUTTON A</button>
    <div>Text</div>
  </template>
</div>

<div id="host-no-delegates-focus-2" class="scroll">
  <template shadowrootmode="open">
    <div style="width: 40em"></div>
    <div>Text</div>
  </template>
</div>

<p>Where host has delegatesFocus true.</p>
<div id="host-delegates-focus" class="scroll">
  <template shadowrootmode="open" shadowrootdelegatesfocus>
    <div style="width: 40em"></div>
    <button id="B">BUTTON B</button>
    <div>Text</div>
  </template>
</div>

<div id="host-delegates-focus-2" class="scroll">
  <template shadowrootmode="open" shadowrootdelegatesfocus>
    <div style="width: 40em"></div>
    <div>Text</div>
  </template>
</div>

<button id="end">END</button>

<script>

promise_test(async () => {
  polyfill_declarative_shadow_dom(document);

  let elements = [
    'start',
    'host-no-delegates-focus/A',
    'host-no-delegates-focus-2',
    'host-delegates-focus/B',
    'end'
  ];

  await assert_focus_navigation_forward(elements);
  elements.reverse();
  await assert_focus_navigation_backward(elements);
}, 'Should focus on scroller across shadow boundaries depending on delegatesFocus value.');

</script>
